backend/requirements.txt
fastapi[standard]
uvicorn
requests
dotenv
evds
tortoise-orm[asyncpg]
aerich
passlib[bcrypt]
fastapi-admin
apscheduler
beautifulsoup4
aiohttp
colorama
